+++ BookingsApi/BookingsApi.Domain/VhoWorkHours.cs
@@ -11,5 +11,17 @@ namespace BookingsApi.Domain
         public TimeSpan? StartTime { get; set; }
         public TimeSpan? EndTime { get; set; }
         public string CreatedBy { get; set; }
+        public System.DayOfWeek SystemDayOfWeek
+        {
+            get
+            {
+                if (DayOfWeekId == 7)
+                {
+                    return System.DayOfWeek.Sunday;
+                }
+
+                return (System.DayOfWeek)DayOfWeekId;
+            }
+        }
     }
 }
